+++ src/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 
+- Fixed boxplot. Remove motion.
 - Sort ordering for groups is more robust.
 - Groups rows can now be deleted through ui.
 - Added version to history to make tracking updates faster.